requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
langchain-google-genai
langchain-community
pypdf
faiss-cpu
langgraph
langchain
python-dotenv
